Move dbpedia-query into utils folder
Ember is a convention-driven framework.  This is experssed in various
folder and filename conventions.  There is tooling to help us find our
way in the project because of these conventions.  It is also good
against bikeshedding arguments such as "should this folder be named
util, or utils" because that was already decided on.

It turns out utils was chosen.  So we will use utils too so our utils
are in the same folder as those generated by the utils class.  Using
the generator does mean we should remove the corresponding file in the
app folder.
